Basic UI Added - GameOver scene - System change commit
When player hits food item, it will either grow or shrink. There is one error that I am getting - when snake hits any fooditem, snake segment is being spawned at same location and it is colliding with snake body and it is triggering the GameOver scene. Need to work on it right now!